requirements.txt
appium-python-client>=3.0.0
selenium>=4.0.0
pytest>=7.0.0
pytest-bdd>=6.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
httpx>=0.25.0
axios
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
aiofiles>=23.2.1
